+++ Coimbra.Editor/GUIDrawers/ManagedFieldDrawer.cs
@@ -108,7 +108,7 @@ namespace Coimbra.Editor
                 Rect valuePosition = position;
                 valuePosition.height = EditorGUI.GetPropertyHeight(systemObject, true);
                 position.height = EditorGUIUtility.singleLineHeight;
-                CoimbraGUIUtility.AdjustPosition(ref position, InspectorArea.Field);
+                EngineUtility.AdjustPosition(ref position, InspectorArea.Field);
 
                 using (GUIContentPool.Pop(out GUIContent typeLabel))
                 {