+++ Coimbra.Editor/ScriptableSettingsEditor.cs
@@ -54,12 +54,12 @@ namespace Coimbra.Editor
         }
 
         /// <summary>
-        /// Helper method that uses <see cref="CoimbraGUIUtility.TryMatchSearch"/> and <see cref="ScriptableSettingsProvider.CurrentSearchContext"/>.
+        /// Helper method that uses <see cref="EngineUtility.TryMatchSearch"/> and <see cref="ScriptableSettingsProvider.CurrentSearchContext"/>.
         /// </summary>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         protected static bool TryMatchSearch(string targetContent)
         {
-            return CoimbraGUIUtility.TryMatchSearch(ScriptableSettingsProvider.CurrentSearchContext, targetContent);
+            return EngineUtility.TryMatchSearch(ScriptableSettingsProvider.CurrentSearchContext, targetContent);
         }
 
         /// <inheritdoc cref="TryMatchSearch(string)"/>
@@ -171,7 +171,7 @@ namespace Coimbra.Editor
                     continue;
                 }
 
-                if (CoimbraGUIUtility.TryMatchSearch(searchContext, iterator.displayName))
+                if (EngineUtility.TryMatchSearch(searchContext, iterator.displayName))
                 {
                     return true;
                 }